conditional_jump fix
