+++ Makefile
@@ -29,18 +29,21 @@ FILES =		destructor.c				\
 		objects/cylinder.c			\
 		objects/factory.c			\
 		objects/light.c				\
-		objects/plane.c				\
+		objects/plane/plane.c			\
+		objects/plane/render.c			\
 		objects/sphere/sphere.c			\
-		objects/sphere/render_sphere.c		\
+		objects/sphere/render.c			\
 		objects/types_factory.c			\
 		objects/updater.c			\
 		parsing/elements.c			\
 		parsing/parser.c			\
 		parsing/utils.c				\
 		render/render.c				\
+		render/specular_reflection.c		\
 		utils/pixel_to_fvector3.c		\
 		utils/fvector3_to_pixel.c		\
 		utils/specular_reflection.c		\
+		utils/clean_buffer.c			\
 
 ifeq ($(shell uname), Darwin)
 FILES += mlx/opengl.c
