+++ src/initializer.c
@@ -36,6 +36,7 @@ t_minirt	*minirt(void)
 		mrt->buffer = malloc(sizeof(t_fvector3) * (WIN_WIDTH * WIN_HEIGHT));
 		if (!mrt->buffer)
 			return (NULL);
+		clean_buffer(mrt);
 		mrt->count = 0;
 		register_all();
 	}
