+++ src/objects/sphere/sphere.c
@@ -28,7 +28,7 @@ t_sphere	*sphere(t_fvector3 position, float diameter, t_rgb color)
 	sp->render = get_render_by_id(SPHERE_ID);
 	sp->pixel_color = 0;
 	sp->pattern = 0;
-	sp->smoothness = 0.0f;
+	sp->smoothness = 1.0f;
 	return (sp);
 }
